+++ src/Detail.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import * as R from 'ramda'
-import { Table, DeleteButton } from './table/Table'
+import { DeleteButton, Table } from './table/Table'
 import { isOneToMany, isManyToMany } from './utils/isType'
 import Field, { getRelSchemaEntry } from './table/Field'
 import {
@@ -18,7 +18,8 @@ import {
 } from './Utils'
 import {
   getActions, getModelAttribute, getField,
-  getDetailFields, getHasIndex, getModelLabel, getFieldLabel, getFieldConditions
+  getDetailFields, getHasIndex, getModelLabel, getFieldLabel,
+  getFieldConditions, getHideable
 } from './utils/schemaGetters'
 import { RecursiveTab } from './Tabs'
 import { getType } from './utils/getType'
@@ -40,6 +41,8 @@ import Input from './form/Input'
 import { Link, Redirect } from 'react-router-dom'
 import '../css/index.css'
 import { inputTypes } from './consts'
+import ReactSVG from 'react-svg'
+import { DeleteDetail } from './delete/DeleteDetail'
 
 const LabelInfoPopover = ({ LabelInfoComponent, fieldLabel }) => (
   <Popover
@@ -51,6 +54,22 @@ const LabelInfoPopover = ({ LabelInfoComponent, fieldLabel }) => (
     labelValue={fieldLabel}
   />
 )
+
+export const HideTableButton = ({ modelName, fieldName, id, hideTable, hideTableChange }) => {
+  const image = hideTable ? 'angle-right' : 'angle-down'
+  return (
+    <ReactSVG
+      src={`/static/img/${image}.svg`}
+      className={`hide-icon-${hideTable ? 'angle-right' : 'angle-down'}`}
+      onClick={() => hideTableChange({ modelName, fieldName, id, hideTable })}
+      svgStyle={{
+        width: '20px',
+        height: '20px',
+        fill: 'black'
+      }}
+    />
+  )
+}
 
 export const DefaultDetailLabel = ({ schema, modelName, fieldName, node, customProps }) => {
   const LabelInfoComponent = R.path(['components', 'labelInfo'], getField(schema, modelName, fieldName))
@@ -224,11 +243,18 @@ export const DefaultDetailTableTitleWrapper = ({ children }) => {
   )
 }
 
-export const DefaultDetailO2MTableTitle = ({ schema, modelName, fieldName, targetInverseFieldName, targetModelName, path, node, user, customProps }) => {
+export const DefaultDetailO2MTableTitle = ({ schema, modelName, fieldName, id, targetInverseFieldName, targetModelName, path, node, user, hideable, hideTable, hideTableChange, customProps }) => {
   const creatable = isCreatable({ schema, modelName: targetModelName, parentNode: node, user, customProps })
 
   return (
     <DefaultDetailTableTitleWrapper>
+      {hideable && <HideTableButton {...{
+        modelName,
+        fieldName,
+        id,
+        hideTable,
+        hideTableChange
+      }}/>}
       <DefaultDetailLabel {...{ schema, modelName, fieldName, node, customProps }} />
       { creatable && <DetailCreateButton {...{
         schema,
@@ -251,12 +277,25 @@ const DefaultDetailM2MTableTitle = ({
   path,
   targetModelName,
   user,
+  hideable,
+  hideTable,
+  hideTableChange,
   customProps
 }) => {
   const editable = isFieldEditable({ schema, modelName, fieldName, node, user, customProps })
+
   return (
     <div style={{ marginBottom: '10px' }}>
-      <h4 className='d-inline'>{getFieldLabel({ schema, modelName, fieldName, node, customProps })}</h4>
+      <h4 className='d-inline'>
+        {hideable && <HideTableButton {...{
+          modelName,
+          fieldName,
+          id,
+          hideTable,
+          hideTableChange
+        }}/>}
+        {getFieldLabel({ schema, modelName, fieldName, node, customProps })}
+      </h4>
       {editable && <div className='pl-2 d-inline'>
         <TableEditButton {...{
           schema,
@@ -314,6 +353,7 @@ export const DefaultDetailTable = ({
   modelStore,
   tooltipData,
   user,
+  tableView,
   modalData,
   customProps
 }) => {
@@ -326,6 +366,9 @@ export const DefaultDetailTable = ({
   const onDelete = R.path(['delete', 'onDetailDelete'], actions)
   const onEditSubmit = R.path(['edit', 'onDetailTableEditSubmit'], actions)
   const type = getType({ schema, modelName, fieldName })
+  const hideTable = R.path(['hideTable', modelName, id, fieldName], tableView)
+  const hideTableChange = R.path(['tableOptions', 'hideTableChange'], actions)
+  const hideable = getHideable(schema, modelName, fieldName)
 
   if (!data) { return <div className='container'>Loading...</div> }
 
@@ -341,11 +384,15 @@ export const DefaultDetailTable = ({
           schema,
           modelName,
           fieldName,
+          id,
           targetInverseFieldName,
           node,
           path,
           targetModelName,
           user,
+          hideable,
+          hideTable,
+          hideTableChange,
           customProps
         }} />
         }
@@ -371,6 +418,8 @@ export const DefaultDetailTable = ({
             }),
             fieldOrder,
             user,
+            tableView,
+            hideTable,
             modalData,
             showHeaders: false,
             customProps
@@ -455,6 +504,9 @@ export const DefaultDetailTable = ({
           path,
           targetModelName,
           user,
+          hideable,
+          hideTable,
+          hideTableChange,
           customProps
         }} /> }
         { skipOverride(ValueOverride) ? null : <DetailValue
@@ -480,6 +532,8 @@ export const DefaultDetailTable = ({
             fieldOrder,
             user,
             showHeaders: false,
+            tableView,
+            hideTable,
             modalData
           }}
         /> }
@@ -513,20 +567,24 @@ const DefaultDetailPageTitle = ({ schema, modelName, node, modalData, user, cust
   const label = getDisplayValue({ schema, modelName, node, customProps })
   const actions = getActions(schema, modelName)
   const onDelete = R.path(['delete', 'onDetailDeleteFromDetailPage'], actions)
+  const onDeleteWarning = R.path(['delete', 'onDeleteWarning'], actions)
+  const modalId = 'confirm-delete-' + modelName
+  const id = R.prop('id', node)
   const HeaderLink = getHasIndex(schema, modelName) ? <Link to={'/' + modelName}>{model}</Link> : model
   return (
     <div><h2 className='d-inline'>{HeaderLink}:<b> {label}</b></h2>
       { isDeletable({ schema, modelName, node, user, customProps }) &&
         <div className='float-right'>
-          <DeleteButton {...{
+          <DeleteButton {...{ modalId, onDeleteWarning, modelName, id }} />
+          <DeleteDetail {...{
             schema,
+            id,
+            modalId,
             modelName,
-            id: node.id,
             onDelete,
-            modalId: 'confirm-delete-' + modelName,
             modalData,
-            customProps }}
-          />
+            customProps
+          }} />
         </div>
       }
     </div>
@@ -547,6 +605,7 @@ export const DetailFields = ({
   modelStore,
   path,
   user,
+  tableView,
   customProps
 }) => {
   if (!node) { return <div className='container'>Loading...</div> }
@@ -567,6 +626,7 @@ export const DetailFields = ({
             return null
           }
           const DetailAttribute = override || DefaultDetailAttribute
+          // same props go into DetailTable & DetailAttribute (even if not used) override gets all same props
           return (
             <DetailAttribute key={`DetailAttribute-${id}-${modelName}-${fieldName}`}
               {...{
@@ -577,10 +637,12 @@ export const DetailFields = ({
                 selectOptions,
                 modelStore,
                 editData,
-                path,
                 tooltipData,
+                modalData,
+                path,
                 id,
                 user,
+                tableView,
                 customProps
               }}
             />
@@ -598,6 +660,7 @@ export const DetailFields = ({
           return null
         }
         const DetailTable = override || DefaultDetailTable
+        // same props go into DetailTable & DetailAttribute (even if not used) override gets all same props
         return (
           <DetailTable
             key={`DetailTable-${id}-${modelName}-${fieldName}`}
@@ -605,15 +668,16 @@ export const DetailFields = ({
               schema,
               modelName,
               fieldName,
+              node,
               selectOptions,
               modelStore,
-              tooltipData,
-              node,
-              modalData,
               editData,
+              tooltipData,
+              modalData,
               path,
               id,
               user,
+              tableView,
               customProps
             }}
           />
@@ -644,6 +708,7 @@ const DefaultDetail = ({
   match,
   tooltipData,
   user,
+  tableView,
   selectOptions,
   modelStore,
   customProps
@@ -705,6 +770,7 @@ const DefaultDetail = ({
             path,
             fields: [],
             user,
+            tableView,
             selectOptions,
             modelStore,
             customProps
@@ -726,6 +792,7 @@ const Detail = ({
   match, // 'match' should be passed in by React by default
   tooltipData,
   user,
+  tableView,
   selectOptions,
   modelStore,
   customProps
@@ -747,6 +814,7 @@ const Detail = ({
         match,
         tooltipData,
         user,
+        tableView,
         selectOptions,
         modelStore,
         customProps