+++ src/input/inputComponent.js
@@ -507,6 +507,7 @@ export const InputCreatableStringSelect = ({
   labelStr,
   id,
   error,
+  className,
   value,
   options,
   onChange,
@@ -515,7 +516,8 @@ export const InputCreatableStringSelect = ({
   customInput,
   customError,
   customLabel
-}) => (
+}) => {
+  return (
   <FormGroup
     labelStr={labelStr}
     htmlFor={id}
@@ -525,6 +527,8 @@ export const InputCreatableStringSelect = ({
     customLabel={customLabel}
   >
     <CreatableSelect
+      className={className}
+      classNamePrefix='select'
       id={id}
       options={options}
       onChange={(selectedOption) => onChange(selectedOption.value)}
@@ -534,4 +538,4 @@ export const InputCreatableStringSelect = ({
       {...customInput}
     />
   </FormGroup>
-)
+)}