+++ CHANGES.md
@@ -8,6 +8,13 @@
 -   Add disabled options dropdown #111
 -   Consolidate filters into one panel #93
 -   Add ability to sort by clicking on field name or header #95
+-   Add table hide feature #109
+-   Give CreatableStringSelect input type className / classNamePrefix #121
+-   Fix broken filter rels #94
+-   Changed Tooltip action model to modelName #122
+-   Fix filter dropdown bug #124
+-   Give filter, hide, and sort icons unique classnames #120
+-   Add pagination #126
 
 ## Version 1.4.0